Extract sections into components
They use contexts,
so they don't need to be
as JSX in the same component.
